Disable UseShellExecute test (#20454)
* Disable test

* Null check

* Explict message for null

+++ src/System.Diagnostics.Process/tests/ProcessStartInfoTests.cs
@@ -944,7 +944,7 @@ namespace System.Diagnostics.Tests
                 FileName = @"http://www.microsoft.com"
             };
 
-            Process.Start(info);
+            Process.Start(info); // Returns null after navigating browser
         }
 
         [ConditionalTheory(nameof(PlatformDetection) + "." + nameof(PlatformDetection.IsNotWindowsNanoServer))] // No Notepad on Nano
@@ -967,6 +967,8 @@ namespace System.Diagnostics.Tests
 
             using (var process = Process.Start(info))
             {
+                Assert.True(process != null, $"Could not start {info.FileName} {info.Arguments} UseShellExecute={info.UseShellExecute}");
+
                 try
                 {
                     process.WaitForInputIdle(); // Give the file a chance to load
@@ -977,7 +979,8 @@ namespace System.Diagnostics.Tests
                 }
                 finally
                 {
-                    process.Kill();
+                    if (process != null && !process.HasExited)
+                        process.Kill();
                 }
             }
         }
@@ -986,6 +989,7 @@ namespace System.Diagnostics.Tests
         [OuterLoop("Launches notepad")]
         [PlatformSpecific(TestPlatforms.Windows)]
         [SkipOnTargetFramework(TargetFrameworkMonikers.Uap, "https://github.com/dotnet/corefx/issues/20204")]
+        [ActiveIssue("https://github.com/dotnet/corefx/issues/20388")]
         public void StartInfo_TextFile_ShellExecute()
         {
             string tempFile = GetTestFilePath() + ".txt";
@@ -1000,6 +1004,8 @@ namespace System.Diagnostics.Tests
 
             using (var process = Process.Start(info))
             {
+                Assert.True(process != null, $"Could not start {info.FileName} UseShellExecute={info.UseShellExecute}");
+
                 try
                 {
                     process.WaitForInputIdle(); // Give the file a chance to load
@@ -1010,7 +1016,8 @@ namespace System.Diagnostics.Tests
                 }
                 finally
                 {
-                    process.Kill();
+                    if (process != null && !process.HasExited)
+                        process.Kill();
                 }
             }
         }
